Remove RecvByteBufAllocator
