+++ servicetalk-tcp-netty-internal/src/main/java/io/servicetalk/tcp/netty/internal/TcpClientChannelInitializer.java
@@ -17,6 +17,7 @@ package io.servicetalk.tcp.netty.internal;
 
 import io.servicetalk.buffer.api.BufferAllocator;
 import io.servicetalk.transport.netty.internal.ChannelInitializer;
+import io.servicetalk.transport.netty.internal.CopyByteBufHandlerChannelInitializer;
 import io.servicetalk.transport.netty.internal.DeferSslHandler;
 import io.servicetalk.transport.netty.internal.IdleTimeoutInitializer;
 import io.servicetalk.transport.netty.internal.SslClientChannelInitializer;
@@ -25,9 +26,6 @@ import io.servicetalk.transport.netty.internal.WireLoggingInitializer;
 import io.netty.channel.Channel;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslHandler;
-
-import static io.servicetalk.transport.netty.internal.PooledRecvByteBufAllocatorInitializers.COPY_HANDLER_INITIALIZER;
-import static io.servicetalk.transport.netty.internal.PooledRecvByteBufAllocatorInitializers.POOLED_RECV_ALLOCATOR_INITIALIZER;
 
 /**
  * {@link ChannelInitializer} for TCP client.
@@ -54,8 +52,7 @@ public class TcpClientChannelInitializer implements ChannelInitializer {
      */
     public TcpClientChannelInitializer(final ReadOnlyTcpClientConfig config, final BufferAllocator alloc,
                                        final boolean deferSslHandler) {
-        ChannelInitializer delegate = ChannelInitializer.defaultInitializer()
-                .andThen(POOLED_RECV_ALLOCATOR_INITIALIZER);
+        ChannelInitializer delegate = ChannelInitializer.defaultInitializer();
 
         if (config.idleTimeoutMs() > 0) {
             delegate = delegate.andThen(new IdleTimeoutInitializer(config.idleTimeoutMs()));
@@ -68,7 +65,7 @@ public class TcpClientChannelInitializer implements ChannelInitializer {
                     config.sslHostnameVerificationPort(), deferSslHandler));
         }
 
-        delegate = delegate.andThen(COPY_HANDLER_INITIALIZER.apply(alloc));
+        delegate = delegate.andThen(new CopyByteBufHandlerChannelInitializer(alloc));
 
         final WireLoggingInitializer wireLoggingInitializer = config.wireLoggingInitializer();
         if (wireLoggingInitializer != null) {
