+++ servicetalk-transport-netty-internal/src/main/java/io/servicetalk/transport/netty/internal/ChannelInitializer.java
@@ -15,7 +15,6 @@
  */
 package io.servicetalk.transport.netty.internal;
 
-import io.netty.channel.AdaptiveRecvByteBufAllocator;
 import io.netty.channel.Channel;
 
 /**
@@ -52,9 +51,6 @@ public interface ChannelInitializer {
      * @return Default initializer for ServiceTalk.
      */
     static ChannelInitializer defaultInitializer() {
-        return channel -> channel.config().setRecvByteBufAllocator(
-                new AdaptiveRecvByteBufAllocator(512, 32768, 65536)
-                        .respectMaybeMoreData(false)
-                        .maxMessagesPerRead(4));
+        return channel -> { /* noop */ };
     }
 }
