+++ servicetalk-transport-netty-internal/src/main/java/io/servicetalk/transport/netty/internal/WrappingSslContext.java
@@ -38,7 +38,7 @@ import javax.annotation.Nullable;
 import javax.net.ssl.SSLEngine;
 import javax.net.ssl.SSLSessionContext;
 
-import static io.servicetalk.transport.netty.internal.PooledRecvByteBufAllocatorInitializers.POOLED_ALLOCATOR;
+import static io.servicetalk.transport.netty.internal.CopyByteBufHandlerChannelInitializer.POOLED_ALLOCATOR;
 import static java.util.Objects.requireNonNull;
 
 final class WrappingSslContext extends SslContext {
